+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Added `Optimove.gamifyWidgetOpen(widgetUrl, userId?, token?)` API to open the Gamify Widget.
 - Android: updated Optimove Android SDK to `7.14.0`.
-- iOS: TBD
+- iOS: updated Optimove iOS SDK to `6.7.0`.
 
 ## 3.1.3